[RENAME] : 컨벤션 통일을 위한 폴더명 변경
