+++ src/pages/QuizComplete/components/InputBox.module.css
