+++ src/pages/QuizComplete/components/InputBox.tsx
@@ -2,7 +2,7 @@ import { FaHeart } from 'react-icons/fa6';
 import { useState } from 'react';
 import { FaRegHeart, FaShareAlt } from 'react-icons/fa';
 import S from './InputBox.module.css';
-import BookMark from '../../../components/ProblemCard/BookMark';
+import BookMark from '../../../components/Card/BookMark';
 import RoundedButton from '../../../components/RoundedButton/RoundedButton';
 
 interface InputBoxProps {
